Updated controllers for db services, add checks and new endpoints; docker for api

+++ TodoListApp.WebApi/Controllers/TodoListController.cs
@@ -88,6 +88,12 @@ public class TodoListController : ControllerBase
             return NotFound();
         }
 
+        var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        if (existing.UserId != userId)
+        {
+            return Forbid();
+        }
+
         await this._todoListService.UpdateListAsync(id, model);
 
         return NoContent();
@@ -100,6 +106,12 @@ public class TodoListController : ControllerBase
         if (existing == null)
         {
             return NotFound();
+        }
+
+        var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        if (existing.UserId != userId)
+        {
+            return Forbid();
         }
 
         await this._todoListService.DeleteListAsync(id);
